Re-organise skiboot code from libstb
Lay out the libstb code in the same way it's laid out in skiboot.

This should make everything else more straightforward later.
