+++ Makefile
@@ -15,8 +15,8 @@ EDK2OBJDIR = backends/edk2-compat
 _EDK2_OBJ =  edk2-svc-read.o edk2-svc-write.o edk2-svc-validate.o edk2-svc-verify.o edk2-svc-generate.o
 EDK2_OBJ = $(patsubst %,$(EDK2OBJDIR)/%, $(_EDK2_OBJ))
 
-SKIBOOTOBJDIR = external/skiboot/
-_SKIBOOT_OBJ = secvar_util.o edk2-compat.o edk2-compat-process.o
+SKIBOOTOBJDIR = external/skiboot/libstb/secvar
+_SKIBOOT_OBJ = secvar_util.o backend/edk2-compat.o backend/edk2-compat-process.o
 SKIBOOT_OBJ = $(patsubst %,$(SKIBOOTOBJDIR)/%, $(_SKIBOOT_OBJ))
 
 OBJ =secvarctl.o  generic.o
