+++ external/skiboot/libstb/secvar/backend/edk2-compat-process.h
@@ -17,18 +17,18 @@
 #include <mbedtls/error.h>
 #include <device.h>
 #include "libstb/crypto/pkcs7/pkcs7.h"
-#include "external/skiboot/include/edk2.h"
+#include "external/skiboot/libstb/secvar/backend/edk2.h"
 #include "opal-api.h"
 #include "../secvar.h"
 #include "../secvar_devtree.h"*/
 
 //added by nick child
-#include "short_types.h"
-#include "opal-api.h"
+#include "external/skiboot/include/short_types.h"
+#include "external/skiboot/include/opal-api.h"
 //#include "external/extraMbedtls/include/pkcs7.h"
 #include "external/skiboot/include/secvar.h"
-#include "external/skiboot/include/edk2.h"
-#include "endian.h"
+#include "external/skiboot/libstb/secvar/backend/edk2.h"
+#include "external/skiboot/include/endian.h"
 
 #define __unused		__attribute__((unused)) //ADDED BY NICK CHILD
