+++ external/skiboot/libstb/secvar/backend/edk2-compat.c
@@ -13,14 +13,14 @@
 /*#include <ccan/endian/endian.h>*/
 /*#include <mbedtls/error.h>*/
 /*#include "libstb/crypto/pkcs7/pkcs7.h"*/
-/*#include "external/skiboot/include/edk2.h"*/
+/*#include "external/skiboot/libstb/secvar/backend/edk2.h"*/
 /*#include "../secvar.h"*/
-#include "external/skiboot/include/edk2-compat-process.h"
+#include "edk2-compat-process.h"
 /*#include "edk2-compat-reset.h"*/
 
 
 //added by Nick Child, not official
-#include "external/skiboot/include/edk2.h"
+#include "edk2.h"
 #include "prlog.h"
 #include "external/skiboot/include/secvar.h"
 #include <stdlib.h>
